third_party/blink/web_tests/wpt_internal/geolocation-api/callback-exception.https.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/blink/public/mojom/geolocation/geolocation_service.mojom.js"></script>
<script src="/gen/third_party/blink/public/mojom/permissions/permission.mojom.js"></script>
<script src="resources/geolocation-mock.js"></script>
<script src="resources/testharness-adapter.js"></script>
</head>
<body>
<script>
setup({allow_uncaught_exception: true});

description("Tests that when an exception is thrown in the success callback, the error callback is not invoked. Note that this test throws an exception which is not caught.");

var mockLatitude = 51.478;
var mockLongitude = -0.166;
var mockAccuracy = 100;

var position;

let gotError = false;
let mock = geolocationMock;
mock.setGeolocationPermission(true);
mock.setGeolocationPosition(mockLatitude,
                            mockLongitude,
                            mockAccuracy);
navigator.geolocation.getCurrentPosition(function(p) {
    position = p;
    shouldBe('position.coords.latitude', 'mockLatitude');
    shouldBe('position.coords.longitude', 'mockLongitude');
    shouldBe('position.coords.accuracy', 'mockAccuracy');

    // Yield to allow for the error callback to be invoked. The timer
    // must be started before the exception is thrown.
    window.setTimeout(assertWeGotException, 0);
    window.onerror = () => { gotError = true; };
    expectError();

    throw new Error('Exception in success callback');
}, function(e) {
    testFailed('Error callback invoked unexpectedly');
    finishJSTest();
});

function assertWeGotException()
{
    assert_true(gotError);
    finishJSTest();
}
</script>
</body>
</html>
